requirements.txt
apache-airflow==2.7.1
pandas==2.1.0
numpy==1.24.3
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
yfinance==0.2.33
scikit-learn==1.3.0
xgboost==2.0.0
pytest==7.4.0
jupyter==1.0.0
matplotlib==3.7.2
seaborn==0.12.2